Convert memoize tests to mocha
